Feature: Configure and start Powergate Daemon
If binary is run with a `--filecoin` flag, the deamon attempts to start a powergate daemon.

Before running the binary, you need to start a localnet instance with docker by running `make localnet-up` from the root direcory.
After docker-compose is done, build and run it with `make && ./bin/space --filecoin`